Add remaining keyword arguments (#21)
Closes #8 

* Add kw to GetCalculatedMetrics

Work towards #8. Instead of using Adobe defaults, chose 100 for limit to avoid needing to page for up to 100 values. Feels like paging for a function like this unnecessary in all but the largest implementations

* Add kwargs to GetReportSuites

* Add kwargs to GetDateRanges

* Add kwargs for GetDimensions

* Fix issue in CRAN check about length

Putting this long of a value in the function signature causes a warning that value will not be printed in documentation PDF. Switch to a NULL and check, which effectively does same thing

* Add kwargs to GetMetrics

* Add kwargs to GetUsers

* Add kwargs for GetSegments