Add Pagination to GET methods (#28)
Adds pagination to GET methods, removes ability for users to page over results themselves. Idea behind this is that an analytics user probably wants all the data anyway and may not be able to write the loop to page over results. Since this package is aiming for a high-level way of interacting with API, simpler to just not let people do whatever they please.